moved breadcrumbs to layout file & fixed mobile search box
